--- a/lib/segment/src/fixtures/sparse_fixtures.rs
+++ b/lib/segment/src/fixtures/sparse_fixtures.rs
@@ -8,7 +8,7 @@ use common::types::PointOffsetType;
 use rand::Rng;
 use sparse::common::sparse_vector::SparseVector;
 use sparse::common::sparse_vector_fixture::random_sparse_vector;
-use sparse::index::inverted_index::inverted_index_immutable_ram::InvertedIndexImmutableRam;
+use sparse::index::inverted_index::inverted_index_compressed_immutable_ram::InvertedIndexImmutableRam;
 use sparse::index::inverted_index::InvertedIndex;
 
 use crate::common::operation_error::OperationResult;
@@ -92,7 +92,7 @@ pub fn fixture_sparse_index_ram<R: Rng + ?Sized>(
     full_scan_threshold: usize,
     data_dir: &Path,
     stopped: &AtomicBool,
-) -> SparseVectorIndex<InvertedIndexImmutableRam> {
+) -> SparseVectorIndex<InvertedIndexImmutableRam<f32>> {
     fixture_sparse_index_ram_from_iter(
         (0..num_vectors).map(|_| random_sparse_vector(rnd, max_dim)),
         full_scan_threshold,
@@ -109,7 +109,7 @@ pub fn fixture_sparse_index_ram_from_iter<P: FnMut()>(
     data_dir: &Path,
     stopped: &AtomicBool,
     progress: impl FnOnce() -> P,
-) -> SparseVectorIndex<InvertedIndexImmutableRam> {
+) -> SparseVectorIndex<InvertedIndexImmutableRam<f32>> {
     let num_vectors = vectors.len();
     let mut sparse_vector_index = fixture_open_sparse_index(
         data_dir,