--- a/lib/segment/src/index/sparse_index/sparse_vector_index.rs
+++ b/lib/segment/src/index/sparse_index/sparse_vector_index.rs
@@ -13,6 +13,7 @@ use itertools::Itertools;
 use sparse::common::scores_memory_pool::ScoresMemoryPool;
 use sparse::common::sparse_vector::SparseVector;
 use sparse::common::types::DimId;
+use sparse::index::inverted_index::inverted_index_ram::InvertedIndexRam;
 use sparse::index::inverted_index::inverted_index_ram_builder::InvertedIndexBuilder;
 use sparse::index::inverted_index::InvertedIndex;
 use sparse::index::migrate::SparseVectorIndexVersion;
@@ -103,12 +104,16 @@ impl<TInvertedIndex: InvertedIndex> SparseVectorIndex<TInvertedIndex> {
                 || (),
             )?;
             (config, inverted_index, indices_tracker)
+        } else if path.read_dir()?.next().is_none() {
+            // Newly created directory - initialize empty inverted index
+            (
+                config,
+                TInvertedIndex::from_ram_index(Cow::Owned(InvertedIndexRam::empty()), path)?,
+                IndicesTracker::default(),
+            )
         } else {
             Self::try_load(path).or_else(|e| {
-                // Avoid noisy warning for newly created segments
-                if vector_storage.borrow().total_vector_count() != 0 {
-                    log::warn!("Failed to load, rebuilding: {}", e.to_string());
-                }
+                log::warn!("Failed to load, rebuilding: {}", e.to_string());
 
                 let (inverted_index, indices_tracker) = Self::build_inverted_index(
                     id_tracker.clone(),
@@ -133,9 +138,7 @@ impl<TInvertedIndex: InvertedIndex> SparseVectorIndex<TInvertedIndex> {
                 // `inverted_index.data` to `inverted_index.dat`.
                 SparseVectorIndexVersion::save(path)?;
 
-                if vector_storage.borrow().total_vector_count() != 0 {
-                    log::info!("Successfully rebuilt");
-                }
+                log::info!("Successfully rebuilt");
 
                 OperationResult::Ok((config, inverted_index, indices_tracker))
             })?